python_api/requirements.txt
pytest
requests
allure-pytest
pyyaml
